Imported upstream version 3.99.pre4

+++ src/fdt/fdt_matrix_ops.cc
@@ -74,7 +74,7 @@ using namespace NEWIMAGE;
 using namespace NEWMAT;
 
 string matf2coordf(string matf){
-  unsigned int pos=matf.rfind("/");
+  size_t pos=matf.rfind("/");
   if(pos!=string::npos)
     matf.replace(pos,1,"/coords_for_");
   else
